Enhanced shopping cart icon
Added a new shopping cart icon without a black background and a "badge" class to framework.css. Modified addtocart/removefromcart functions in the ProductsController to store the total number of items in the cart into session and added the count as a badge shown next to the cart icon when there are 1+ products in cart.